split lark
